requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=2.0.0
requests>=2.31.0
pydantic>=2.4.0
structlog>=23.1.0
pytz>=2023.3
python-multipart>=0.0.6
boto3==1.34.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
loguru==0.7.2
finnhub-python==2.4.19 